Fix double-click on list seperator bug
Double-clicking on the divider between the main menu list and entry list
would cause an ever-growing main menu list until the app is reloaded.

This magic incantation fixes it. I still have no idea what it does; certainly doesn't
appear to be "auto" in any of my understandings of the word so can't
even understand what the original intention of the mystery magic code
might be. Presumably it's some poorly documented backbone stuff.